added makefile

+++ ProgC/Lab_12_1_2/main.py
@@ -5,18 +5,18 @@ lib = ctypes.CDLL('./mylib.dll')
 # int get_new_mas(int *src, int src_len, int *dst, int *dst_len)
 _get_new_mas = lib.get_new_mas
 
-_get_new_mas.argtypes = (ctypes.POINTER(ctypes.c_double), ctypes.c_int, ctypes.POINTER(ctypes.c_double), ctypes.POINTER(ctypes.c_int))
+_get_new_mas.argtypes = (ctypes.POINTER(ctypes.c_int), ctypes.c_int, ctypes.POINTER(ctypes.c_int), ctypes.POINTER(ctypes.c_int))
 _get_new_mas.restype = ctypes.c_int
 
 def get_new_mas(nums):
   n = len(nums)
-  arr = (ctypes.c_double * n)(*nums)
+  arr = (ctypes.c_int * n)(*nums)
 
   n_res = ctypes.c_int(0)
 
   rc = _get_new_mas(arr, n, None, n_res)
   if rc:
-    res = (ctypes.c_double * n_res.value)()
+    res = (ctypes.c_int * n_res.value)()
 
     rc = _get_new_mas(arr, n, res, n_res)
